+++ mobile/android/android-components/components/support/webextensions/src/main/java/mozilla/components/support/webextensions/WebExtensionSupport.kt
@@ -294,12 +294,14 @@ object WebExtensionSupport {
 
                 override fun onInstallPermissionRequest(
                     extension: WebExtension,
-                    onPermissionsGranted: ((Boolean) -> Unit),
+                    permissions: List<String>,
+                    onPermissionsGranted: (Boolean) -> Unit,
                 ) {
                     store.dispatch(
                         WebExtensionAction.UpdatePromptRequestWebExtensionAction(
                             WebExtensionPromptRequest.AfterInstallation.Permissions.Required(
                                 extension,
+                                permissions,
                                 onPermissionsGranted,
                             ),
                         ),